Used size_t instead of long in a few instances of reinterpret_cast

+++ src/gui/auxiliary/event/handler.cpp
@@ -331,7 +331,7 @@ void thandler::handle_event(const SDL_Event& event)
 			break;
 
 		case TIMER_EVENT:
-			execute_timer(reinterpret_cast<long>(event.user.data1));
+			execute_timer(reinterpret_cast<size_t>(event.user.data1));
 			break;
 
 		case CLOSE_WINDOW_EVENT:
